:art: refactor : Service를 확장성을 고려하여 인터페이스와 구현체로 분리
